Move `watch_objs` API call across modules (no changes, just a move)
